console warnings fixed, confirmation to delete education/ experience added, profile component added, main information, education and experience profile components added

+++ client/src/actions/profile.js
@@ -184,39 +184,43 @@ export const addEducation = (formData, history) => async (dispatch) => {
 
 // Delete Experience
 export const deleteExperience = (id) => async (dispatch) => {
-  try {
-    const res = await axios.delete(`/api/profile/experience/${id}`);
+  if (window.confirm('Are you sure? Experience will be deleted')) {
+    try {
+      const res = await axios.delete(`/api/profile/experience/${id}`);
 
-    dispatch({
-      type: UPDATE_PROFILE,
-      payload: res.data,
-    });
+      dispatch({
+        type: UPDATE_PROFILE,
+        payload: res.data,
+      });
 
-    dispatch(setAlert('Experience Deleted', 'success'));
-  } catch (err) {
-    dispatch({
-      type: PROFILE_ERROR,
-      payload: { msg: err.response.statusText, status: err.response.status },
-    });
+      dispatch(setAlert('Experience Deleted', 'success'));
+    } catch (err) {
+      dispatch({
+        type: PROFILE_ERROR,
+        payload: { msg: err.response.statusText, status: err.response.status },
+      });
+    }
   }
 };
 
 // Delete Education
 export const deleteEducation = (id) => async (dispatch) => {
-  try {
-    const res = await axios.delete(`/api/profile/education/${id}`);
+  if (window.confirm('Are you sure? Education will be deleted')) {
+    try {
+      const res = await axios.delete(`/api/profile/education/${id}`);
 
-    dispatch({
-      type: UPDATE_PROFILE,
-      payload: res.data,
-    });
+      dispatch({
+        type: UPDATE_PROFILE,
+        payload: res.data,
+      });
 
-    dispatch(setAlert('Education Deleted', 'success'));
-  } catch (err) {
-    dispatch({
-      type: PROFILE_ERROR,
-      payload: { msg: err.response.statusText, status: err.response.status },
-    });
+      dispatch(setAlert('Education Deleted', 'success'));
+    } catch (err) {
+      dispatch({
+        type: PROFILE_ERROR,
+        payload: { msg: err.response.statusText, status: err.response.status },
+      });
+    }
   }
 };
 
@@ -224,7 +228,7 @@ export const deleteEducation = (id) => async (dispatch) => {
 export const deleteAccount = () => async (dispatch) => {
   if (window.confirm('Are you sure? This can NOT be undone!')) {
     try {
-      const res = await axios.delete('/api/profile/');
+      await axios.delete('/api/profile/');
 
       dispatch({ type: CLEAR_PROFILE });
       dispatch({ type: ACCOUNT_DELETED });
